requirements.txt
streamlit
torch
torchvision
numpy
pandas
scikit-learn
Pillow
openai
python-dotenv
watchdog